fix deleted elements on duplicate

+++ src/app/components/application/application-form/application-form.component.ts
@@ -472,7 +472,7 @@ export class ApplicationFormComponent implements OnInit {
     data.forEach(parameter => {
       if (parameter.status === 'pendingCreation' || parameter.status === 'pendingModify') {
           parameter.application=this.applicationToEdit} //If is new, you need the application link
-          if(parameter.status === 'pendingCreation'){
+          if(parameter.status === 'pendingCreation' || parameter.new){
             parameter._links=null;
             parameter.id = null;
           }
@@ -480,7 +480,7 @@ export class ApplicationFormComponent implements OnInit {
           promises.push(new Promise((resolve, reject) => { this.applicationParameterService.save(parameter).subscribe((resp) => { resolve(true) }) }));
 
       
-      if(parameter.status === 'pendingDelete' && parameter._links) {
+      if(parameter.status === 'pendingDelete' && parameter._links && !parameter.new) {
         // parameterToDelete.push(parameter) 
         promises.push(new Promise((resolve, reject) => { this.applicationParameterService.remove(parameter).subscribe((resp) => { resolve(true) }) }));
 
@@ -582,6 +582,8 @@ export class ApplicationFormComponent implements OnInit {
 
   getAllRowsRoles(data: any[] )
   {
+    const promises: Promise<any>[] = [];
+
     let dataChanged = false;
     let rolesModified = [];
     let rolesToPut = [];
@@ -589,7 +591,8 @@ export class ApplicationFormComponent implements OnInit {
       
       if(role.status!== 'pendingDelete') {
         if (role.status === 'pendingModify') {
-          rolesModified.push(role) 
+          promises.push(new Promise((resolve, reject) => { this.roleService.update(role).subscribe((resp) => { resolve(true) }) }));
+          if(role.new){ dataChanged = true; }
         }
         else if(role.status === 'pendingCreation') {
           dataChanged = true;
@@ -600,29 +603,16 @@ export class ApplicationFormComponent implements OnInit {
 
     });
 
-    console.log(rolesModified);
-    this.updateRoles(rolesModified, rolesToPut, dataChanged);
-
-  }
-
-  updateRoles(rolesModified: Role[], rolesToPut: Role[], dataChanged: boolean)
-  {
-    const promises: Promise<any>[] = [];
-    rolesModified.forEach(role => {
-      promises.push(new Promise((resolve, reject) => { this.roleService.update(role).subscribe((resp) => { resolve(true) }) }));
+    Promise.all(promises).then(() => {   
+      if(dataChanged)
+      {
+        let url=this.applicationToEdit._links.availableRoles.href.split('{', 1)[0];
+        this.utils.updateUriList(url,rolesToPut, this.dataUpdatedEventRoles);
+      }
+      else{ this.dataUpdatedEventRoles.next(true) }
     });
 
-     Promise.all(promises).then(() => {   
-        if(dataChanged)
-        {
-          let url=this.applicationToEdit._links.availableRoles.href.split('{', 1)[0];
-          this.utils.updateUriList(url,rolesToPut, this.dataUpdatedEventRoles);
-        }
-        else{ this.dataUpdatedEventRoles.next(true) }
-      });
   }
-  
- 
 
   // ******** Background ******** //
 
@@ -683,13 +673,9 @@ export class ApplicationFormComponent implements OnInit {
             // backgroundsToCreate.push(background) 
             promises.push(new Promise((resolve, reject) => { this.applicationBackgroundService.save(background).subscribe((resp) => { resolve(true) }) }));
           }
-        
-
         }
-
-        
       }
-      if(background.status === 'pendingDelete' ) {
+      if(background.status === 'pendingDelete' && !background.new ) {
         // backgroundsToDelete.push(background) 
         promises.push(new Promise((resolve, reject) => { this.applicationBackgroundService.remove(background).subscribe((resp) => { resolve(true) }) }));
 
@@ -750,10 +736,14 @@ export class ApplicationFormComponent implements OnInit {
     let dataChanged = false;
     let treesModified = [];
     let treesToPut = [];
+    const promises: Promise<any>[] = [];
+
     data.forEach(tree => {
       if(tree.status!== 'pendingDelete') {
         if (tree.status === 'pendingModify') {
-          treesModified.push(tree);
+          // treesModified.push(tree);
+          if(tree.new){ dataChanged = true; }
+          promises.push(new Promise((resolve, reject) => { this.treeService.update(tree).subscribe((resp) => { resolve(true) }) }));
         }
         else if (tree.status === 'pendingCreation'){
           dataChanged = true;
@@ -764,23 +754,15 @@ export class ApplicationFormComponent implements OnInit {
     });
 
     console.log(treesModified);
-    this.updateTrees(treesModified, treesToPut, dataChanged);
-  }
 
-  updateTrees(treesModified: Tree[], treesToPut: Tree[], dataChanged: boolean)
-  {
-    const promises: Promise<any>[] = [];
-    treesModified.forEach(tree => {
-      promises.push(new Promise((resolve, reject) => { this.treeService.update(tree).subscribe((resp) => { resolve(true) }) }));
+    Promise.all(promises).then(() => {
+      if(dataChanged)
+      {
+        let url=this.applicationToEdit._links.trees.href.split('{', 1)[0];
+        this.utils.updateUriList(url,treesToPut,this.dataUpdatedEventTree)
+      }
+      else {this.dataUpdatedEventTree.next(true); }
     });
-      Promise.all(promises).then(() => {
-        if(dataChanged)
-        {
-          let url=this.applicationToEdit._links.trees.href.split('{', 1)[0];
-          this.utils.updateUriList(url,treesToPut,this.dataUpdatedEventTree)
-        }
-        else {this.dataUpdatedEventTree.next(true); }
-      });
 
   }
